+++ test/executor_test.rb
@@ -530,7 +530,7 @@ module Dynflow
           world.plan(Support::CodeWorkflowExample::IncomingIssues, issues_data)
         end
 
-        let(:manager) { Executors::Parallel::FlowManager.new execution_plan, execution_plan.run_flow }
+        let(:manager) { Director::FlowManager.new execution_plan, execution_plan.run_flow }
 
         def assert_next_steps(expected_next_step_ids, finished_step_id = nil, success = true)
           if finished_step_id
